requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
Pillow
firebase-admin
gunicorn
dj-database-url
psycopg2-binary
whitenoise
